Rename: auth 폴더명 -> Authentication로 변경 (+ 경로수정)
