+++ src/pages/Authentication/RegisterPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegisterForm from '../../components/Auth/Register';
+import RegisterForm from '../../components/Authentication/Register';
 
 const RegisterPage: React.FC = () => {
   return (
