+++ src/styles/Authentication/LoginRegisterSC.ts
